Add MPS parser
