+++ smps/parsers/__init__.py
@@ -1,3 +1,4 @@
 from .CoreParser import CoreParser
+from .MpsParser import MpsParser
 from .StochParser import StochParser
 from .TimeParser import TimeParser
