+++ smps/parsers/tests/test_MpsParser.py
@@ -1,0 +1,6 @@
+"""
+Historically, the CoreParser came first. The MpsParser was split off only later.
+As such, most tests are present in test_CoreParser.py, rather than this file.
+"""
+
+# TODO
